\name{ClassPrediction-package}
\Rdversion{1.1}
\alias{ClassPrediction-package}
\alias{ClassPrediction}
\docType{package}
\title{
What the package does (short line)
Provides tools to help with class prediction
}
\description{
Provides tools to help with class prediction
}
\details{
\tabular{ll}{
Package: \tab ClassPrediction\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-27\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Kevin R. Coombes <kcoombes@mdanderson.org>

Maintainer: Kevin R. Coombes <kcoombes@mdanderson.org>
}
\references{
David Goldberg. "Genetic Algorithms in Search, Optimization and Machine
Learning," Addison-Wesley, 1989.
}
\keyword{ package }
\keyword{optimize}
\keyword{models}
\keyword{multivariate}
\keyword{classif}
\seealso{
}
\examples{
# not here yet
}
