\name{ClassPrediction-package}
\Rdversion{1.1}
\alias{ClassPrediction-package}
\alias{ClassPrediction}
\docType{package}
\title{The ClassPrediction Bundle}
\description{
  This class provides tools to help with class prediction. In its current
  form, it is a shell used to bundle various other packages that actually
  do all the work.
}
\details{
  \tabular{ll}{
    Package: \tab ClassPrediction\cr
    Type: \tab Package\cr
    Version: \tab 2.10.2\cr
    Date: \tab 2012-05-01\cr
    License: \tab Artistic-2.0\cr
    LazyLoad: \tab yes\cr
  }
}
\author{
  Kevin R. Coombes \email{kcoombes@mdanderson.org}
}
\seealso{
  \code{\link{GenAlg-class}},
  \code{\link{CrossValidate-class}}.
}
\keyword{package}

