\name{ClassPrediction-package}
\Rdversion{1.1}
\alias{ClassPrediction-package}
\alias{ClassPrediction}
\docType{package}
\title{
The ClassPrediction Bundle
}
\description{
This class provides tools to help with class prediction. In its current
form, it is a shell used to bundle various other packages that actually
do all the work.
}
\details{
\tabular{ll}{
Package: \tab ClassPrediction\cr
Type: \tab Package\cr
Version: \tab 2.10\cr
Date: \tab 2010-01-06\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
}
See the components, \code{\link{GenAlg-class}} and \code{\link{CrossVal-class}}.
}
\author{
Kevin R. Coombes <kcoombes@mdanderson.org>

Maintainer: Kevin R. Coombes <kcoombes@mdanderson.org>
}
\keyword{ package }
\keyword{optimize}
\keyword{models}
\keyword{multivariate}
\keyword{classif}
\seealso{
  \code{\link{GenAlg-class}}, \code{\link{CrossVal-class}}.
}
\examples{
# see the component packages.
}
